idpa/requirements.txt
# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.11.7
pdf2image==1.17.0
pytesseract==0.3.13

# Enhanced PDF Processing and Layout Analysis
layoutparser==0.3.4
detectron2==0.6
easyocr==1.7.0
sympy==1.12
opencv-python==4.8.1.78
mathpix-python==0.1.0

# Text Processing and NLP
nltk==3.9.1
sentence-transformers==5.0.0
torch>=2.2.0

# Vector Storage and Search
faiss-cpu==1.11.0
numpy==1.26.4

# LLM Integration
openai==1.97.0

# Data Processing
pandas==2.1.4
scikit-learn==1.7.0

# CLI and Interface
typer==0.12.5
rich==14.0.0
jupyter==1.1.1
ipywidgets==8.1.5

# Utilities
python-dotenv==1.1.1
tqdm==4.67.0
requests==2.32.4

# Development and Testing
pytest==8.3.3
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1

# Image Processing (for OCR and Enhanced Extraction)
Pillow==11.3.0 